backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0